// src/types/enums.ts
export enum eCallType {
    Incoming = "Incoming",
    Outgoing = "Outgoing",
    Missed = "Missed",
    Unanswered = "Unanswered"
}

export enum eFilterHeadings {
    Ascending = "Ascending",
    Descending = "Descending",
    None = "None"
}
export enum eAssessment {
    Bad =  "Bad",
    Good = "Good",
    Excellent = "Excellent",
}